Default to potentially measuring all commits (instead of just --first-parent path)
When you measure all commits the measurements might be taken alternatingly
between two long branches that were both merged to master eventually and
the measurements might give different values for each branch. This results
in a graph with lots of jitter up and down between whatever the typical
metric value was for the first and the second branch.

Because git history is non-linear and can have two different paths along
the tree (that eventually merges), it's very hard to render a graph showing
the "value of metric FOO in this repo at time X" because at time X there
might literally be two (or more) different values for the metric FOO (one
for each branch that existed at time X).

In many cases the best to workaround this by adjusting the metric script
or metric command, for example for a metric that measures commit count
you get jitter if you do:

  git log --oneline %(sha1)s --since="1 month ago" | wc -l

and you get more accurate values with no graph jitter if you filter the
entire repo each time and count the number of commits between the commit
timestamp of the %(sha1)s commit and one month earlier.

However, one drawback of the above approach is that commits that were made
on a branch in for example May and then merged to master in June, will be
counted as "commits in May" with treepath: "all". In some cases you might
want those commits booked as coming in to master in June instead and you
get that behavior instead if you set treepath: "first-parent" (this also
assumes that every merge to master always had the old master as its first
parent, which is often but not always the case).

Both of these two workarounds to the "it's not possible to graph a metric
across non-linear git history as a single series" are non-optimal but if
you look at the git history of a particular repo you can sometimes see that
one or the other workaround is more suitable.